chore: move brave rewards types
