+++ components/brave_wallet_ui/common/slices/endpoints/rewards.endpoints.test.ts
@@ -8,13 +8,15 @@ import { renderHook } from '@testing-library/react-hooks'
 // queries
 import { useGetRewardsInfoQuery } from '../api.slice'
 
+// constants
+import { WalletStatus } from '../../../constants/types'
+
 // utils
 import {
   createMockStore,
   renderHookOptionsWithMockStore
 } from '../../../utils/test-utils'
 import { BraveRewardsProxyOverrides } from '../../async/__mocks__/brave_rewards_api_proxy'
-import { WalletStatus } from '../../async/brave_rewards_api_proxy'
 
 const mockedRewardsData: BraveRewardsProxyOverrides = {
   externalWallet: {
