+++ components/brave_wallet_ui/components/desktop/views/portfolio/portfolio-overview.tsx
@@ -27,9 +27,9 @@ import {
 import {
   BraveWallet,
   UserAssetInfoType,
-  WalletRoutes
+  WalletRoutes,
+  WalletStatus
 } from '../../../../constants/types'
-import { WalletStatus } from '../../../../common/async/brave_rewards_api_proxy'
 import { emptyRewardsInfo } from '../../../../common/async/base-query-cache'
 
 // actions
